Developer

+++ src/main/java/com/prosubject/prosubject/backend/apirest/controller/ProfesorController.java
@@ -1,8 +1,10 @@
 package com.prosubject.prosubject.backend.apirest.controller;
 
+import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.TimeUnit;
 
 import org.junit.Assert;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -98,7 +100,7 @@ public class ProfesorController {
 			response.put("mensaje", "El DNI ya esta en uso por un profesor");
 			return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
 		} else if (emailsEnuso.contains(prof.getEmail())) {
-			response.put("mensaje", "El email ya no se encuentra en uso");
+			response.put("mensaje", "El email ya se encuentra en uso");
 			return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
 		} else if (usersEnuso.contains(prof.getUserAccount().getUsername())) {
 			response.put("mensaje", "El nombre de usuario ya esta en uso");
@@ -106,14 +108,16 @@ public class ProfesorController {
 		} else {
 
 			try {
-				if (file!=null) {
+				if (file != null) {
 					if (!file.getContentType().equalsIgnoreCase("application/pdf")) {
 						response.put("mensaje", "El expediente no tiene formato pdf");
 						return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
 
 					} else {
-						DBFile borrarPdf = this.profesorService.findOne(id).getExpendiente();
-						this.dBFileStorageService.delete(borrarPdf);
+						if (p.getExpendiente() != null) {
+							DBFile borrarPdf = this.profesorService.findOne(id).getExpendiente();
+							this.dBFileStorageService.delete(borrarPdf);
+						}
 						DBFile dbFile = this.dBFileStorageService.storeFile(file);
 						prof.setExpendiente(dbFile);
 					}
@@ -173,15 +177,135 @@ public class ProfesorController {
 		return new ResponseEntity<Profesor>(profesorModificado, HttpStatus.OK);
 	}
 
-	@GetMapping("/validacionExpedientePendiente")
-	public List<Profesor> profesoresExpedientePendiete() {
-		return this.profesorService.profesoresExpedientePendiete();
-
-	}
+	//----------Tarifa Premium----------//
 	
-	@GetMapping("/profesoresTarifaPremium")
-	public List<Profesor> profesoresTarifaPremium() {
-		return this.profesorService.profesoresExpedientePendiete();
-	}
+		@GetMapping("/pagoTarifaPremium/{id}")
+		public ResponseEntity<?> inscribirPremium(@PathVariable Long id) {
+			Map<String, Object> response = new HashMap<String, Object>();
+			Profesor profesorModificado = null;
+			Profesor prof = this.profesorService.findOne(id);
+			Assert.assertNotNull(prof);
+			
+		if(prof.getTarifaPremium()==Boolean.TRUE) {
+			//Comprueba si ya lleva 30 dias con el premium
+			if(numeroDiasEntreDosFechas(prof.getFechaPagoPremium(),new Date())<30) {
+				Integer diasPremium = 30 - (numeroDiasEntreDosFechas(prof.getFechaPagoPremium(),new Date()));
+				response.put("mensaje", "El profesor aun posee "+String.valueOf(diasPremium)+"dias de plan premium");
+				return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+			}
+			
+			
+			else if(numeroDiasEntreDosFechas(prof.getFechaPagoPremium(),new Date())==30){
+				response.put("mensaje", "El plan premium del profesor acaba hoy");
+				return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+			}
+		}
+			else {
+			try {
+				prof.setTarifaPremium(Boolean.TRUE);
+				prof.setFechaPagoPremium(new Date());
+				profesorModificado = this.profesorService.save(prof);
+			} catch (DataAccessException e) {
+				response.put("mensaje", "Error al realizar el update en la base de datos");
+				response.put("error", e.getMessage().concat(": ").concat(e.getMostSpecificCause().getMessage()));
+				return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+			}
+			}
+		return new ResponseEntity<Profesor>(profesorModificado, HttpStatus.OK);
+		}
+		
+		
+		
+		//------Comprueba cuantos dias de premium le quedan al profesor------//
+		
+		@GetMapping("/comprobarDiasPremium/{id}")
+		public ResponseEntity<?> diasPremium(@PathVariable Long id) {
+			Map<String, Object> response = new HashMap<String, Object>();
+			Profesor prof = this.profesorService.findOne(id);
+			Assert.assertNotNull(prof);
+			
+			if(prof.getTarifaPremium()==Boolean.TRUE) {
+				//Comprueba si ya lleva 30 dias con el premium
+				if(numeroDiasEntreDosFechas(prof.getFechaPagoPremium(),new Date())<30) {
+					Integer diasPremium = 30 - (numeroDiasEntreDosFechas(prof.getFechaPagoPremium(),new Date()));
+					response.put("mensaje", "El profesor aun posee "+String.valueOf(diasPremium)+"dias de plan premium");
+					return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+				}
+				
+				else if(numeroDiasEntreDosFechas(prof.getFechaPagoPremium(),new Date())==30){
+					response.put("mensaje", "El plan premium del profesor acaba hoy");
+					return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+				}
+			}
+			/*else {
+			try {
+				
+			} catch (DataAccessException e) {
+				response.put("mensaje", "Error al realizar la consulta en la base de datos");
+				response.put("error", e.getMessage().concat(": ").concat(e.getMostSpecificCause().getMessage()));
+				return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+			}
+	*/
+			return new ResponseEntity<Profesor>(prof, HttpStatus.OK);
+		}
+
+		//--------------------------//
+		
+		
+		@GetMapping("/validacionExpedientePendiente")
+		public List<Profesor> profesoresExpedientePendiete() {
+			return this.profesorService.profesoresExpedientePendiete();
+
+		}
+		
+		@GetMapping("/profesoresTarifaPremium")
+		public List<Profesor> profesoresTarifaPremium() {
+			return this.profesorService.profesoresExpedientePendiete();
+		}
+		
+		
+		///-------Auxiliar
+		public static int numeroDiasEntreDosFechas(Date fecha1, Date fecha2){
+		     long startTime = fecha1.getTime();
+		     long endTime = fecha2.getTime();
+		     long diffTime = endTime - startTime;
+		     return (int)TimeUnit.DAYS.convert(diffTime, TimeUnit.MILLISECONDS);
+		}
+		
+		
+		
+		@PutMapping("/peticionBorrar/{profesorId}")
+		public ResponseEntity<?> modificarHorario(@PathVariable Long profesorId) throws Exception {
+			Map<String, Object> response = new HashMap<String, Object>();
+			Profesor profesor = this.profesorService.findOne(profesorId);
+			
+			if(profesor==null) {
+				response.put("mensaje",	 "El profesor con ID: ".concat(profesorId.toString()).concat(" no existe"));
+				return new ResponseEntity<Map<String, Object>>(response, HttpStatus.NOT_FOUND);
+			}
+			if(profesor.getDerechoOlvidado()==true) {
+				response.put("mensaje",	 "El profesor ya ha solicitado  la peticion para ser olvidado ");
+				return new ResponseEntity<Map<String, Object>>(response, HttpStatus.NOT_FOUND);
+			}
+			if(this.profesorService.profesorTieneAlumno(profesor.getId())!=true) {
+				response.put("mensaje",	 "No se pudo realizar la petición porque aún tiene horarios activos.");
+				return new ResponseEntity<Map<String, Object>>(response, HttpStatus.NOT_FOUND);
+			}
+			
+			try {
+				profesor  =this.profesorService.peticionBorrar(profesor);
+			}catch(DataAccessException e) {
+				response.put("mensaje", "Error al realizar el insert en la base de datos");
+				response.put("error", e.getMessage().concat(": ").concat(e.getMostSpecificCause().getMessage()));
+				return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR); 
+					
+			}
+			
+			return new ResponseEntity<Profesor>(profesor,HttpStatus.CREATED); 
+			
+			
+			
+		}
+
 
 }
